Move default badge definitions into separate module
