+++ indigo_social/badges.py
@@ -5,7 +5,6 @@ from django.contrib.auth.models import User
 
 from pinax.badges.base import Badge, BadgeAwarded, BadgeDetail
 from pinax.badges.registry import badges
-from allauth.account.signals import user_signed_up
 from indigo_api.models import Country
 from indigo_app.models import Editor
 
@@ -237,52 +236,3 @@ def editor_countries_changed(sender, instance, action, reverse, model, pk_set, *
 def country_saved(sender, instance, created, raw, **kwargs):
     if created:
         CountryBadge.create_for_country(instance)
-
-
-# ------------------------------------------------------------------------
-# Badge definitions
-
-
-class ContributorBadge(PermissionBadge):
-    slug = 'contributor'
-    name = 'Contributor'
-    group_name = name + ' Badge'
-    description = 'Can view work details'
-    permissions = ('indigo_api.add_annotation', 'indigo_api.change_annotation', 'indigo_api.delete_annotation',
-                   'indigo_api.add_task')
-
-
-class DrafterBadge(PermissionBadge):
-    slug = 'drafter'
-    name = 'Drafter'
-    group_name = name + ' Badge'
-    description = 'Can create new works and edit the details of existing works'
-    permissions = ('indigo_api.add_work', 'indigo_api.change_work',
-                   'indigo_api.add_document', 'indigo_api.change_document',
-                   'indigo_api.add_amendment', 'indigo_api.change_amendment', 'indigo_api.delete_amendment',
-                   # required when restoring a document version
-                   'reversion.change_version',
-                   'indigo_api.change_task', 'indigo_api.submit_task', 'indigo_api.reopen_task',
-                   'indigo_api.add_workflow', 'indigo_api.change_workflow')
-
-
-class SeniorDrafterBadge(PermissionBadge):
-    slug = 'senior-drafter'
-    name = 'Senior Drafter'
-    group_name = name + ' Badge'
-    description = 'Can review work tasks and delete documents and works'
-    permissions = ('indigo_api.delete_work', 'indigo_api.review_work',
-                   'indigo_api.review_document', 'indigo_api.delete_document', 'indigo_api.publish_document',
-                   'indigo_api.cancel_task', 'indigo_api.unsubmit_task', 'indigo_api.close_task',
-                   'indigo_api.close_workflow', 'indigo_api.delete_workflow')
-
-
-badges.register(ContributorBadge)
-badges.register(DrafterBadge)
-badges.register(SeniorDrafterBadge)
-
-
-# when a user signs up, grant them the contributor badge immediately
-@receiver(user_signed_up, sender=User)
-def grant_contributor_new_user(sender, request, user, **kwargs):
-    badges.registry['contributor'].award(user=user)
